+++ lpg2/tests/golden/minimal/dart/minimalprs.dart
@@ -54,6 +54,9 @@ import 'minimalsym.dart';
     static const bool  BACKTRACK = false;
     bool  getBacktrack() { return BACKTRACK; }
 
+    static const bool  GLR = false;
+    bool  isGLR() { return GLR; }
+
     int  getStartSymbol() { return lhs(0); }
     bool  isValidForParser()  { return minimalsym.isValidForParser; }